+++ src/test/java/org/folio/api/SearchBrowseSuiteIT.java
@@ -11,14 +11,14 @@ import static org.folio.search.model.types.ResourceType.INSTANCE_CALL_NUMBER;
 import static org.folio.search.model.types.ResourceType.INSTANCE_CLASSIFICATION;
 import static org.folio.search.model.types.ResourceType.INSTANCE_CONTRIBUTOR;
 import static org.folio.search.model.types.ResourceType.INSTANCE_SUBJECT;
+import static org.folio.search.model.types.ResourceType.LINKED_DATA_AUTHORITY;
 import static org.folio.search.model.types.ResourceType.LINKED_DATA_HUB;
-import static org.folio.search.model.types.ResourceType.LINKED_DATA_INSTANCE;
 import static org.folio.search.model.types.ResourceType.LINKED_DATA_WORK;
 import static org.folio.support.TestConstants.TENANT_ID;
 import static org.folio.support.testdata.SharedTestDataManager.LockManager;
 import static org.folio.support.testdata.SharedTestDataManager.instancesCount;
+import static org.folio.support.testdata.SharedTestDataManager.linkedDataAuthoritiesCount;
 import static org.folio.support.testdata.SharedTestDataManager.linkedDataHubsCount;
-import static org.folio.support.testdata.SharedTestDataManager.linkedDataInstancesCount;
 import static org.folio.support.testdata.SharedTestDataManager.linkedDataWorksCount;
 import static org.folio.support.testdata.SharedTestDataManager.loadAll;
 
@@ -38,8 +38,8 @@ import org.folio.api.search.SearchByEmptyValuesIT;
 import org.folio.api.search.SearchHoldingsIT;
 import org.folio.api.search.SearchInstanceIT;
 import org.folio.api.search.SearchItemIT;
+import org.folio.api.search.SearchLinkedDataAuthorityIT;
 import org.folio.api.search.SearchLinkedDataHubIT;
-import org.folio.api.search.SearchLinkedDataInstanceIT;
 import org.folio.api.search.SearchLinkedDataWorkIT;
 import org.folio.api.search.SortAuthorityIT;
 import org.folio.api.search.SortInstanceIT;
@@ -111,7 +111,7 @@ class SearchBrowseSuiteIT extends BaseIntegrationTest {
     awaitIndexedResourceCounts(INSTANCE_CLASSIFICATION, TENANT_ID, EXPECTED_CLASSIFICATION_COUNT);
     awaitIndexedResourceCounts(INSTANCE_CONTRIBUTOR, TENANT_ID, EXPECTED_CONTRIBUTOR_COUNT);
     awaitIndexedResourceCounts(INSTANCE_SUBJECT, TENANT_ID, EXPECTED_SUBJECT_COUNT);
-    awaitIndexedResourceCounts(LINKED_DATA_INSTANCE, TENANT_ID, linkedDataInstancesCount());
+    awaitIndexedResourceCounts(LINKED_DATA_AUTHORITY, TENANT_ID, linkedDataAuthoritiesCount());
     awaitIndexedResourceCounts(LINKED_DATA_WORK, TENANT_ID, linkedDataWorksCount());
     awaitIndexedResourceCounts(LINKED_DATA_HUB, TENANT_ID, linkedDataHubsCount());
   }
@@ -162,10 +162,10 @@ class SearchBrowseSuiteIT extends BaseIntegrationTest {
   class SearchItem extends SearchItemIT { }
 
   @Nested
-  class SearchLinkedDataHub extends SearchLinkedDataHubIT { }
+  class SearchLinkedDataAuthority extends SearchLinkedDataAuthorityIT { }
 
   @Nested
-  class SearchLinkedDataInstance extends SearchLinkedDataInstanceIT { }
+  class SearchLinkedDataHub extends SearchLinkedDataHubIT { }
 
   @Nested
   class SearchLinkedDataWork extends SearchLinkedDataWorkIT { }